src/stacks/ingest/st_poller/lambda/requirements.txt
python-twitter
aws-lambda-powertools